kube: split service watcher to new file.
It's complicated enough that thinking about it separately helps.
